Break newsched into multiple files
